requirements.txt
langchain
crewai
langchain-google-genai
langchain-groq
python-dotenv
praw
resend
Markdown
